+++ src/mlpack/config.hpp
@@ -148,4 +148,15 @@
   #undef MLPACK_USE_SYSTEM_HTTPLIB
 #endif
 
+// MLPACK_DISABLE_REMOTE_DATASET_CACHE disables the download cache.
+// This feature is enabled by default.
+#if !defined(MLPACK_DISABLE_REMOTE_DATASET_CACHE)
+  // #define MLPACK_DISABLE_REMOTE_DATASET_CACHE
+#endif
+
+// MLPACK_REMOTE_DATASET_CACHE_DIR can be set to override the default cache
+// directory.  If not set, the default is $HOME/.mlpack/cache/ on POSIX and
+// %APPDATA%\mlpack\cache\ on Windows.
+// #define MLPACK_REMOTE_DATASET_CACHE_DIR "/custom/cache/path/"
+
 #endif